getBookStatistics() 리팩토링 - 동일한 기능 DB에서 처리(JPQL)

+++ kopring-study/src/main/kotlin/com/yeongenn/kopringstudy/kopring/domain/dto/BookStatResponse.kt
@@ -4,7 +4,8 @@ import com.yeongenn.kopringstudy.kopring.domain.entity.BookType
 
 data class BookStatResponse(
     val type: BookType,
-    val count: Int,
+    //val count: Int,
+    val count: Long,
 ) {
 //    fun plusOne() {
 //        this.count += 1
